twitter-responder/requirements.txt
boto3
python-twitter
